Start actually drawing things

+++ Font.h
@@ -1,8 +1,10 @@
 #pragma once
 
+#include <inttypes.h>
+
 typedef struct Font Font;
 
 Font *FontNew(const char *path, float size);
 int FontDelete(Font *font);
-int DrawText(const char *text, float x, float y, Font *font);
+int DrawText(const char *text, int32_t x, int32_t y, Font *font);
 int GetTextSize(const char *text, int *width, int *height, Font *font);
